# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SupaWeb Agency API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  technologies?: string[];
  category: string;
  /** @nullable */
  liveUrl?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  featured?: boolean;
  published?: boolean;
  createdAt: string;
}

export interface ProjectInput {
  title: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  technologies?: string[];
  category: string;
  /** @nullable */
  liveUrl?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  featured?: boolean;
  published?: boolean;
}

export interface ProjectUpdate {
  title?: string;
  description?: string;
  /** @nullable */
  imageUrl?: string | null;
  technologies?: string[];
  category?: string;
  /** @nullable */
  liveUrl?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  featured?: boolean;
  published?: boolean;
}

export interface Service {
  id: number;
  title: string;
  description: string;
  icon: string;
  features?: string[];
  sortOrder: number;
}

export interface ServiceInput {
  title: string;
  description: string;
  icon: string;
  features?: string[];
  sortOrder?: number;
}

export interface ServiceUpdate {
  title?: string;
  description?: string;
  icon?: string;
  features?: string[];
  sortOrder?: number;
}

export interface Testimonial {
  id: number;
  name: string;
  role: string;
  company: string;
  content: string;
  /** @nullable */
  avatarUrl?: string | null;
  rating?: number;
  createdAt: string;
}

export interface TestimonialInput {
  name: string;
  role: string;
  company: string;
  content: string;
  /** @nullable */
  avatarUrl?: string | null;
  rating?: number;
}

export interface TestimonialUpdate {
  name?: string;
  role?: string;
  company?: string;
  content?: string;
  /** @nullable */
  avatarUrl?: string | null;
  rating?: number;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  subject?: string | null;
  message: string;
  read?: boolean;
  createdAt: string;
}

export interface ContactInput {
  name: string;
  email: string;
  /** @nullable */
  subject?: string | null;
  message: string;
}

export interface AdminUser {
  id: number;
  email: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface PublicStats {
  projectsCount: number;
  servicesCount: number;
  testimonialsCount: number;
  yearsActive: number;
}

export interface AdminStats {
  projectsCount: number;
  servicesCount: number;
  testimonialsCount: number;
  messagesCount: number;
  unreadMessagesCount: number;
  featuredProjectsCount?: number;
  publishedProjectsCount?: number;
}

export type ListProjectsParams = {
  category?: string;
};