# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SupaWeb Agency API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Portfolio projects
  - name: services
    description: Agency services
  - name: testimonials
    description: Client testimonials
  - name: contact
    description: Contact form
  - name: auth
    description: Admin authentication
  - name: admin
    description: Admin management
  - name: stats
    description: Statistics and analytics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  # --- Public endpoints ---

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List published projects
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a single project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found

  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all services
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"

  /testimonials:
    get:
      operationId: listTestimonials
      tags: [testimonials]
      summary: List published testimonials
      responses:
        "200":
          description: List of testimonials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Testimonial"

  /stats:
    get:
      operationId: getPublicStats
      tags: [stats]
      summary: Get public agency stats
      responses:
        "200":
          description: Agency statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicStats"

  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit a contact message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Message submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"
        "400":
          description: Validation error

  # --- Auth endpoints ---

  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Invalid credentials

  /auth/logout:
    post:
      operationId: adminLogout
      tags: [auth]
      summary: Admin logout
      responses:
        "200":
          description: Logged out

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current admin session
      responses:
        "200":
          description: Current admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authenticated

  # --- Admin project management ---

  /admin/projects:
    get:
      operationId: adminListProjects
      tags: [admin]
      summary: List all projects (admin)
      responses:
        "200":
          description: All projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [admin]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /admin/projects/{id}:
    patch:
      operationId: updateProject
      tags: [admin]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [admin]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- Admin service management ---

  /admin/services:
    post:
      operationId: createService
      tags: [admin]
      summary: Create a service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"

  /admin/services/{id}:
    patch:
      operationId: updateService
      tags: [admin]
      summary: Update a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [admin]
      summary: Delete a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- Admin testimonial management ---

  /admin/testimonials:
    post:
      operationId: createTestimonial
      tags: [admin]
      summary: Create a testimonial
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestimonialInput"
      responses:
        "201":
          description: Created testimonial
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"

  /admin/testimonials/{id}:
    patch:
      operationId: updateTestimonial
      tags: [admin]
      summary: Update a testimonial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestimonialUpdate"
      responses:
        "200":
          description: Updated testimonial
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
    delete:
      operationId: deleteTestimonial
      tags: [admin]
      summary: Delete a testimonial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- Admin contact messages ---

  /admin/messages:
    get:
      operationId: listMessages
      tags: [admin]
      summary: List all contact messages
      responses:
        "200":
          description: All contact messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactMessage"

  /admin/messages/{id}:
    delete:
      operationId: deleteMessage
      tags: [admin]
      summary: Delete a contact message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- Admin dashboard stats ---

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard statistics
      responses:
        "200":
          description: Admin statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Project:
      type: object
      required: [id, title, description, category, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        technologies:
          type: array
          items:
            type: string
        category:
          type: string
        liveUrl:
          type: ["string", "null"]
        githubUrl:
          type: ["string", "null"]
        featured:
          type: boolean
        published:
          type: boolean
        createdAt:
          type: string

    ProjectInput:
      type: object
      required: [title, description, category]
      properties:
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        technologies:
          type: array
          items:
            type: string
        category:
          type: string
        liveUrl:
          type: ["string", "null"]
        githubUrl:
          type: ["string", "null"]
        featured:
          type: boolean
        published:
          type: boolean

    ProjectUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        technologies:
          type: array
          items:
            type: string
        category:
          type: string
        liveUrl:
          type: ["string", "null"]
        githubUrl:
          type: ["string", "null"]
        featured:
          type: boolean
        published:
          type: boolean

    Service:
      type: object
      required: [id, title, description, icon, sortOrder]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        features:
          type: array
          items:
            type: string
        sortOrder:
          type: integer

    ServiceInput:
      type: object
      required: [title, description, icon]
      properties:
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        features:
          type: array
          items:
            type: string
        sortOrder:
          type: integer

    ServiceUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        features:
          type: array
          items:
            type: string
        sortOrder:
          type: integer

    Testimonial:
      type: object
      required: [id, name, role, company, content, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        company:
          type: string
        content:
          type: string
        avatarUrl:
          type: ["string", "null"]
        rating:
          type: integer
        createdAt:
          type: string

    TestimonialInput:
      type: object
      required: [name, role, company, content]
      properties:
        name:
          type: string
        role:
          type: string
        company:
          type: string
        content:
          type: string
        avatarUrl:
          type: ["string", "null"]
        rating:
          type: integer

    TestimonialUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        company:
          type: string
        content:
          type: string
        avatarUrl:
          type: ["string", "null"]
        rating:
          type: integer

    ContactMessage:
      type: object
      required: [id, name, email, message, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        subject:
          type: ["string", "null"]
        message:
          type: string
        read:
          type: boolean
        createdAt:
          type: string

    ContactInput:
      type: object
      required: [name, email, message]
      properties:
        name:
          type: string
        email:
          type: string
        subject:
          type: ["string", "null"]
        message:
          type: string

    AdminUser:
      type: object
      required: [id, email]
      properties:
        id:
          type: integer
        email:
          type: string

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    PublicStats:
      type: object
      required: [projectsCount, servicesCount, testimonialsCount, yearsActive]
      properties:
        projectsCount:
          type: integer
        servicesCount:
          type: integer
        testimonialsCount:
          type: integer
        yearsActive:
          type: integer

    AdminStats:
      type: object
      required: [projectsCount, servicesCount, testimonialsCount, messagesCount, unreadMessagesCount]
      properties:
        projectsCount:
          type: integer
        servicesCount:
          type: integer
        testimonialsCount:
          type: integer
        messagesCount:
          type: integer
        unreadMessagesCount:
          type: integer
        featuredProjectsCount:
          type: integer
        publishedProjectsCount:
          type: integer