# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SupaWeb Agency API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  projectsCount: number;
  servicesCount: number;
  testimonialsCount: number;
  messagesCount: number;
  unreadMessagesCount: number;
  featuredProjectsCount?: number;
  publishedProjectsCount?: number;
}