# lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SupaWeb Agency API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInput {
  name: string;
  email: string;
  /** @nullable */
  subject?: string | null;
  message: string;
}