# lib/api-zod/src/generated/types/projectUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SupaWeb Agency API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectUpdate {
  title?: string;
  description?: string;
  /** @nullable */
  imageUrl?: string | null;
  technologies?: string[];
  category?: string;
  /** @nullable */
  liveUrl?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  featured?: boolean;
  published?: boolean;
}